split out kubeclient and testpod
